refactored Printer into separate file
